queryselector(all), eventlistener 등 함수개선
