--- config.js
@@ -1,2 +1,0 @@
-export const USER_EMAIL = 'test@codeit.com';
-export const USER_PASSWORD = 'codeit101';
