+++ index.html
@@ -29,7 +29,7 @@
         <a href="index.html">
           <img src="/assets/images/index/logo.svg" alt="홈으로 연결된 Linkbrary 로고" />
         </a>
-        <a class="cta cta-short" href="/signin.html">
+        <a class="cta cta-short" href="/pages/login/signin.html">
           <span>로그인</span>
         </a>
       </nav>
