+++ pages/common/config/regex.js
