+++ pages/common/helper.js
@@ -1,0 +1,12 @@
+export function qs(selector, scope = document) {
+  if (!selector) throw 'no selector';
+  return scope.querySelector(selector);
+}
+
+export function qsAll(selector, scope = document) {
+  if (!selector) throw 'no selector';
+  return Array.from(scope.querySelectorAll(selector));
+}
+export function on(target, eventName, handler) {
+  target.addEventListener(eventName, handler);
+}
