+++ pages/login/login.js
@@ -1,7 +1,7 @@
-import LoginController from './controller/loginController.js';
-import LoginView from './view/loginView.js';
-import LoginService from './service/loginService.js';
-import User from './domain/user.js';
+import LoginController from './loginController.js';
+import LoginView from './loginView.js';
+import LoginService from './loginService.js';
+import user from '../domain/user.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   const accessToken = localStorage.getItem('accessToken');
@@ -10,7 +10,6 @@ document.addEventListener('DOMContentLoaded', () => {
     return;
   }
   const loginView = new LoginView();
-  const user = new User();
   const loginService = new LoginService(loginView, user);
   new LoginController(loginView, user, loginService);
 });
