+++ pages/login/signin.html
@@ -2,7 +2,7 @@
 <html lang="ko">
   <head>
     <meta charset="utf-8" />
-    <title>회원가입</title>
+    <title>로그인</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css" />
 
@@ -12,7 +12,7 @@
     <header>
       <div class="header_container">
         <a class="header_container_logo" href="/"><img src="/assets/images/index/logo.svg" alt="홈으로 연결된 Linkbrary 로고" /></a>
-        <p class="header_container__p">회원이 아니신가요?<a class="header_container__a" href="./signin.html">로그인 하기</a></p>
+        <p class="header_container__p">회원이 아니신가요?<a class="header_container__a" href="../signup/signup.html">회원 가입하기</a></p>
       </div>
     </header>
 
@@ -30,26 +30,18 @@
             <button class="eye-button eye-on" type="button" tabindex="-1"></button>
             <p class="error_msg"></p>
           </div>
-          <div class="sign_box_input sign-password">
-            <label for="password">비밀번호 확인</label>
-            <input type="password" name="password_chk" id="password_chk" />
-            <button class="eye-button eye-on" type="button" tabindex="-1"></button>
-            <p class="error_msg"></p>
-          </div>
         </div>
-        <input type="submit" value="회원가입" />
+        <input type="submit" value="로그인" />
       </form>
 
       <div class="singin_oauth">
         <p>소셜 로그인</p>
         <div class="sign_oauth__icon">
-          <a href="https://www.google.com" target="_blank"><img src="/assets/images/signin/google.png" alt="구글홈페이지이동아이콘" /></a>
-          <a href="https://www.kakaocorp.com/page" target="_blank"
-            ><img src="/assets/images/signin/kakao.png" alt="카카오홈페이지이동아이콘"
-          /></a>
+          <a href="https://www.google.com" target="_blank"><img src="/assets/images/signin/google.png" alt="" /></a>
+          <a href="https://www.kakaocorp.com/page" target="_blank"><img src="/assets/images/signin/kakao.png" alt="" /></a>
         </div>
       </div>
     </div>
-    <script type="module" src="/assets/js/sign/signUp.js"></script>
+    <script type="module" src="./login.js"></script>
   </body>
 </html>
