+++ pages/signup/signUp.js
@@ -1,7 +1,7 @@
-import SignUpController from './controller/signUpController.js';
-import SignUpView from './view/signUpView.js';
-import SignUpService from './service/SignupService.js';
-import User from './domain/user.js';
+import SignUpController from './signUpController.js';
+import SignUpView from './signUpView.js';
+import SignUpService from './SignupService.js';
+import user from '../domain/user.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   const accessToken = localStorage.getItem('accessToken');
@@ -11,7 +11,6 @@ document.addEventListener('DOMContentLoaded', () => {
   }
 
   const signUpView = new SignUpView();
-  const user = new User();
   const signUpService = new SignUpService(signUpView, user);
   new SignUpController(signUpView, user, signUpService);
 });
